--- controllers/CONTROLLERS_README.txt
@@ -1,2 +1,0 @@
-# This directory is where we will write all of our controllers
-  -Sal Camara Sept 12 1:06pm
