--- entities/Professor.java
@@ -1,50 +1,0 @@
-package com.banner.bannerApplication.entities;
-
-import javax.persistence.*;
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-public class Professor {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(unique = true)
-    private Long id;
-
-    @OneToMany(mappedBy = "professor")
-    private Set<Section> section = new HashSet<>();
-
-    @Column(unique = false)
-    private String firstName;
-
-    @Column(unique = false)
-    private String lastName;
-
-    public Professor(){}
-    public Professor(String firstName, String lastName){
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public Set<Section> getSection(){
-        return section;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-}
