--- entities/Rooms.java
@@ -1,41 +1,0 @@
-package com.banner.bannerApplication.entities;
-
-import javax.persistence.*;
-
-@Entity
-public class Rooms {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String roomSize;
-
-    @Column(unique = true)
-    private String roomNumber;
-
-    public Long getId(){
-        return id;
-    }
-
-    public String getRoomSize() {
-        return roomSize;
-    }
-
-    public void setRoomSize(String roomSize) {
-
-        this.roomSize = roomSize;
-    }
-
-    public String getRoomNumber() {
-
-        return roomNumber;
-    }
-
-    public void setRoomNumber(String roomNumber) {
-
-        this.roomNumber = roomNumber;
-    }
-
-}
-
